requirements.txt
streamlit
google-generativeai
youtube-transcript-api
yt-dlp
python-dotenv
toml